requirements.txt
absl-py==1.0.0
addict==2.4.0
beautifulsoup4==4.10.0
brotlipy @ file:///home/conda/feedstock_root/build_artifacts/brotlipy_1648854164153/work
cachetools==5.0.0
certifi==2022.6.15
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1644853463426/work
click==8.1.2
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1649035231366/work
cycler==0.11.0
docker-pycreds==0.4.0
filelock==3.6.0
fonttools==4.28.5
gdown==4.4.0
gensim==4.2.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.2
google-auth-oauthlib==0.4.6
grpcio==1.44.0
huggingface-hub==0.8.1
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1642433548627/work
importlib-metadata==4.11.3
inplace-abn==1.1.1.dev5+g2d86dbf.d20220504
joblib==1.1.0
kiwisolver==1.4.2
Markdown==3.3.6
matplotlib==3.5.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mmcv-full==1.3.16
nltk==3.7
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1634106693478/work
oauthlib==3.2.0
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
opencv-python==4.6.0.66
packaging==21.3
pathtools==0.1.2
pickle5 @ file:///home/conda/feedstock_root/build_artifacts/pickle5_1637450421899/work
Pillow==6.2.1
promise==2.3
protobuf==3.20.0
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1643496850550/work
pyparsing==3.0.9
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1648857264451/work
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.7.25
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1641580202195/work
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
sentence-transformers==2.2.2
sentencepiece==0.1.96
sentry-sdk==1.5.8
setproctitle==1.2.2
shortuuid==1.0.8
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smart-open==6.0.0
smmap==5.0.0
soupsieve==2.3.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.11.0
torchsummary==1.5.1
torchvision==0.12.0
tqdm==4.64.0
transformers==4.21.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1644850595256/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1647489083693/work
wandb==0.12.11
Werkzeug==2.1.1
yapf==0.32.0
yaspin==2.1.0
zipp==3.8.0
